# backend/tests/fixtures/openapi/breaking.json
{
  "components": {
    "schemas": {
...
      "Body_upload_novel_api_novels_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_novel_api_novels_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportStatusResponse": {
        "description": "小说导入状态响应（用于前端轮询进度）",
        "properties": {
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "导入任务 ID",
            "title": "Job Id"
          },
          "message": {
            "default": "",
            "description": "阶段描述信息",
            "title": "Message",
            "type": "string"
          },
          "novel_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "小说 ID（创建前可能为 null）",
            "title": "Novel Id"
          },
          "percent": {
            "description": "进度百分比 0-100",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percent",
            "type": "integer"
          },
          "stage": {
            "description": "当前阶段: uploading / detecting / parsing / saving / ready / failed / error",
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "stage",
          "percent"
        ],
        "title": "ImportStatusResponse",
        "type": "object"
      },
...
      "NovelBulkDeleteRequest": {
        "description": "批量删除请求；限制单次规模，避免误操作和超长事务。",
        "properties": {
          "novel_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Novel Ids",
            "type": "array"
          }
        },
        "required": [
          "novel_ids"
        ],
        "title": "NovelBulkDeleteRequest",
        "type": "object"
      },
      "NovelBulkDeleteResponse": {
        "description": "批量删除结果。未找到或无权访问的 ID 统一计入 skipped_ids。",
        "properties": {
          "deleted_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Deleted Ids",
            "type": "array"
          },
          "skipped_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Skipped Ids",
            "type": "array"
          }
        },
        "required": [
          "deleted_ids",
          "skipped_ids"
        ],
        "title": "NovelBulkDeleteResponse",
        "type": "object"
      },
...
      "NovelUploadResponse": {
        "description": "小说上传结果响应（含处理状态和统计信息）",
        "properties": {
          "chapter_count": {
            "default": 0,
            "title": "Chapter Count",
            "type": "integer"
          },
          "id": {
            "description": "导入任务 job_id（兼容字段，用于轮询）",
            "title": "Id",
            "type": "integer"
          },
          "job_id": {
            "description": "导入任务 ID",
            "title": "Job Id",
            "type": "integer"
          },
          "message": {
            "default": "文件已上传，正在处理中...",
            "title": "Message",
            "type": "string"
          },
          "novel_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "小说 ID（任务完成前可能为 null）",
            "title": "Novel Id"
          },
          "status": {
            "description": "处理状态: pending / importing / ready / failed",
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "word_count": {
            "default": 0,
            "title": "Word Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "job_id",
          "title",
          "status"
        ],
        "title": "NovelUploadResponse",
        "type": "object"
      },
...
      "ReadingProgressUpdate": {
        "description": "更新阅读进度请求",
        "properties": {
          "chapter_id": {
            "description": "当前阅读到的章节ID",
            "title": "Chapter Id",
            "type": "integer"
          },
          "progress_percent": {
            "description": "阅读进度百分比 0-100",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress Percent",
            "type": "number"
          }
        },
        "required": [
          "chapter_id",
          "progress_percent"
        ],
        "title": "ReadingProgressUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI 辅助小说创作与理解平台",
    "title": "NovelMind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/novels": {
      "get": {
        "description": "获取小说列表（分页 + 搜索）",
        "operationId": "list_novels_api_novels_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "搜索标题或作者",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "搜索标题或作者",
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Novels",
        "tags": [
          "小说管理"
        ]
      }
    },
    "/api/novels/bulk": {
      "delete": {
        "description": "批量删除当前用户可管理的小说。",
        "operationId": "bulk_delete_novels_api_novels_bulk_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovelBulkDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NovelBulkDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Bulk Delete Novels",
        "tags": [
          "小说管理"
        ]
      }
    },
    "/api/novels/import-jobs/{job_id}": {
      "get": {
        "description": "按 job_id 查询导入进度（上传后应轮询此接口，而不是 novel_id）。",
        "operationId": "get_import_job_status_api_novels_import_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Import Job Status",
        "tags": [
          "小说管理"
        ]
      }
    },
    "/api/novels/upload": {
      "post": {
        "description": "上传并解析小说 TXT 文件（支持多编码自动检测和大文件）。\n\n需要登录认证。\n\n处理流程:\n1. 在请求内读完文件字节（避免 BackgroundTask 时 UploadFile 已关闭）\n2. 创建导入任务（ImportJob）\n3. 独立 DB 会话后台处理\n4. 返回 job_id 供前端轮询 GET /novels/import-jobs/{job_id}",
        "operationId": "upload_novel_api_novels_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_novel_api_novels_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NovelUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Novel",
        "tags": [
          "小说管理"
        ]
      }
    },
    "/api/novels/{novel_id}": {
      "delete": {
        "description": "删除小说（仅限所有者或超级用户）",
        "operationId": "delete_novel_api_novels__novel_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Novel",
        "tags": [
          "小说管理"
        ]
      },
      "get": {
        "description": "获取小说详情（含章节列表）",
        "operationId": "get_novel_api_novels__novel_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NovelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Novel",
        "tags": [
          "小说管理"
        ]
      },
      "patch": {
        "description": "更新小说名称等元信息（仅限所有者或超级用户）。",
        "operationId": "update_novel_api_novels__novel_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovelUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NovelListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Novel",
        "tags": [
          "小说管理"
        ]
      }
    },
    "/api/novels/{novel_id}/chapters": {
      "get": {
        "description": "获取小说章节列表（不含完整正文，避免大 payload）",
        "operationId": "list_chapters_api_novels__novel_id__chapters_get",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChapterSummaryResponse"
                  },
                  "title": "Response List Chapters Api Novels  Novel Id  Chapters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Chapters",
        "tags": [
          "小说管理"
        ]
      }
    },
    "/api/novels/{novel_id}/chapters/{chapter_id}": {
      "get": {
        "description": "获取章节内容",
        "operationId": "get_chapter_api_novels__novel_id__chapters__chapter_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Chapter",
        "tags": [
          "小说管理"
        ]
      }
    },
...
    "/api/novels/{novel_id}/import-cancel": {
      "post": {
        "description": "取消正在运行的导入任务",
        "operationId": "cancel_import_api_novels__novel_id__import_cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Import",
        "tags": [
          "小说管理"
        ]
      }
    },
    "/api/novels/{novel_id}/import-retry": {
      "post": {
        "description": "重试失败的导入任务",
        "operationId": "retry_import_api_novels__novel_id__import_retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry Import",
        "tags": [
          "小说管理"
        ]
      }
    },
    "/api/novels/{novel_id}/import-status": {
      "get": {
        "description": "获取小说导入进度状态（前端轮询用；优先使用 /import-jobs/{job_id}）",
        "operationId": "get_import_status_api_novels__novel_id__import_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Import Status",
        "tags": [
          "小说管理"
        ]
      }
    },
...
    "/api/novels/{novel_id}/progress": {
      "patch": {
        "description": "更新阅读进度",
        "operationId": "update_progress_api_novels__novel_id__progress_patch",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadingProgressUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Progress",
        "tags": [
          "小说管理"
        ]
      }
    },
...
    "/api/search/novels/{novel_id}": {
      "post": {
        "description": "小说内混合搜索。\n\n结合 BM25 全文搜索和向量搜索，加权融合排序。\n\n可选认证：有 token 则检查所有权，无 token 则允许访问。",
        "operationId": "novel_search_api_search_novels__novel_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Novel Search",
        "tags": [
          "搜索"
        ]
      }
    },
...
  }
}